# openapi.yaml
openapi: "3.1.0"
info:
  title: ChronoGuard API
  description: |
    DST-aware datetime validation, resolution, and conversion API.

    **Need an API key?** [Get one free on the homepage](/) — no credit card required. Then click **Authorize** above and paste your key.
  version: "1.1.0"
servers:
  - url: https://chronoguard-api-production.up.railway.app
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - apiKey: []
components:
  securitySchemes:
    apiKey:
      type: apiKey
      name: x-api-key
      in: header
paths:
  /v1/datetime/validate:
    post:
      summary: Validate a local datetime
      description: Check if a local datetime is valid, invalid (DST gap), or ambiguous (DST overlap) in the given timezone
      tags: [datetime]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [local_datetime, time_zone]
              properties:
                local_datetime:
                  type: string
                  example: "2026-03-08T02:30:00"
                time_zone:
                  type: string
                  example: "America/New_York"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [valid, invalid, ambiguous]
                  reason_code:
                    type: string
                    enum: [DST_GAP, DST_OVERLAP, INVALID_TIMEZONE]
                  message:
                    type: string
                  suggested_fixes:
                    type: array
                    items:
                      type: object
                      properties:
                        strategy:
                          type: string
                        local_datetime:
                          type: string
                  possible_instants:
                    type: array
                    items:
                      type: object
                      properties:
                        offset:
                          type: string
                        instant_utc:
                          type: string

  /v1/datetime/resolve:
    post:
      summary: Resolve a local datetime to UTC
      description: Resolve an ambiguous or invalid local datetime to a UTC instant using the specified policy
      tags: [datetime]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [local_datetime, time_zone]
              properties:
                local_datetime:
                  type: string
                  example: "2026-11-01T01:30:00"
                time_zone:
                  type: string
                  example: "America/New_York"
                resolution_policy:
                  type: object
                  properties:
                    ambiguous:
                      type: string
                      enum: [earlier, later, reject]
                      default: earlier
                    invalid:
                      type: string
                      enum: [next_valid_time, previous_valid_time, reject]
                      default: next_valid_time
      responses:
        "200":
          description: Resolved UTC instant
          content:
            application/json:
              schema:
                type: object
                properties:
                  instant_utc:
                    type: string
                  offset:
                    type: string

  /v1/datetime/convert:
    post:
      summary: Convert UTC to local time
      description: Convert a UTC instant to a local datetime in the target timezone
      tags: [datetime]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instant_utc, target_time_zone]
              properties:
                instant_utc:
                  type: string
                  example: "2026-06-15T15:00:00Z"
                target_time_zone:
                  type: string
                  example: "Europe/London"
      responses:
        "200":
          description: Converted local datetime
          content:
            application/json:
              schema:
                type: object
                properties:
                  local_datetime:
                    type: string
                  offset:
                    type: string
                  time_zone:
                    type: string

  /v1/datetime/batch:
    post:
      summary: Batch process multiple datetime operations
      description: Process up to 100 validate, resolve, and convert operations in a single request
      tags: [datetime]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  maxItems: 100
                  items:
                    type: object
                    required: [operation]
                    properties:
                      operation:
                        type: string
                        enum: [validate, resolve, convert]
                      local_datetime:
                        type: string
                      time_zone:
                        type: string
                      instant_utc:
                        type: string
                      target_time_zone:
                        type: string
                      resolution_policy:
                        type: object
                        properties:
                          ambiguous:
                            type: string
                            enum: [earlier, later, reject]
                          invalid:
                            type: string
                            enum: [next_valid_time, previous_valid_time, reject]
      responses:
        "200":
          description: Batch results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        index:
                          type: integer
                        operation:
                          type: string
                        success:
                          type: boolean
                        data:
                          type: object
                        error:
                          type: object
                  total:
                    type: integer
                  succeeded:
                    type: integer
                  failed:
                    type: integer

  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok